[코드 및 파일 추가] 게임 시작 버튼으로 서버 IP를 입력하여 연결하도록 구현, 플레이어가 2명일 때 인게임 씬으로 넘어가도록 구현, 전체적인 코드 점검, .dds 이미지 파일 교체

+++ PrisonBreaker/BilboardMesh.cpp
@@ -48,12 +48,12 @@ const XMUINT2& CBilboardMesh::GetCellCount() const
 	return m_CellCount;
 }
 
-void CBilboardMesh::SetCellIndex(UINT CellIndex)
+void CBilboardMesh::SetCellIndex(float CellIndex)
 {
 	m_CellIndex = CellIndex;
 }
 
-UINT CBilboardMesh::GetCellIndex() const
+float CBilboardMesh::GetCellIndex() const
 {
 	return m_CellIndex;
 }
